Added Universal Module Definition
