+++ demo/noajax-fr.html
@@ -5,7 +5,7 @@
     <title>Outdated Browser</title>
     <meta name="description" content="A time saving tool for developers. It detects outdated browsers and advises users to upgrade to a new version.">
     <!-- Styles -->
-    <link rel="stylesheet" href="../outdatedbrowser/outdatedbrowser.min.css">
+    <link rel="stylesheet" href="../dist/outdatedbrowser.css">
     <link href='http://fonts.googleapis.com/css?family=Open+Sans' rel='stylesheet' type='text/css'>
     <style type="text/css">
         body {
@@ -42,7 +42,7 @@
     </div>
 
     <!-- javascript includes -->
-    <script src="../outdatedbrowser/outdatedbrowser.min.js"></script>
+    <script src="../dist/outdatedbrowser.js"></script>
 
     <!-- plugin call -->
     <script>
