+++ gulpfile.js
@@ -1,0 +1,2 @@
+require('coffee-script/register');
+require('./gulpfile.coffee');
