+++ package.json
@@ -1,0 +1,30 @@
+{
+  "name": "outdated-browser",
+  "version": "1.1.0",
+  "description": "A time saving tool for developers. It detects outdated browsers and advises users to upgrade to a new version.",
+  "main": "./src/outdatedbrowser.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/burocratik/outdated-browser"
+  },
+  "author": "burocratik",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/burocratik/outdated-browser/issues"
+  },
+  "homepage": "https://github.com/burocratik/outdated-browser",
+  "devDependencies": {
+    "browser-sync": "^1.9.0",
+    "coffee-script": "^1.8.0",
+    "gulp": "^3.8.10",
+    "gulp-csso": "^0.2.9",
+    "gulp-jshint": "^1.9.0",
+    "gulp-load-plugins": "^0.8.0",
+    "gulp-size": "^1.2.0",
+    "gulp-uglify": "^1.0.2",
+    "jshint-stylish": "^1.0.0"
+  }
+}
